Changed all Double embedding stuff to Float.

+++ Sources/CleverBird/embeddings/typealiases.swift
@@ -1,6 +1,6 @@
 //  Created by B.T. Franklin on 8/31/23
 
 public typealias Document = String
-public typealias Vector = [Double]
-public typealias Similarity = Double
+public typealias Vector = [Float]
+public typealias Similarity = Float
 public typealias SimilarityFunction = ([Vector], Vector) -> [Similarity]
